Update spacing in mod.txt
Spacing in previous mod.txt was very weird and hard to look at, so I put in spaces where they made sense and took out the ones that didn't so it was much easier to read. Doesn't change any part of the mod simply makes mod.txt easier to look at.